STOR-8 refactor to model dir
